Refactor code per review
